refactor hunter lab
