+++ lymui/lab.c
@@ -48,51 +48,6 @@ Lab *getLabFromXyz(Xyz *xyz) {
     return lab;
 }
 
-/////// HunterLab calculation
-
-/**
- * @brief Get Ka Kb return the Ka Kb value based on the Xyz value (should exposed ?)
- * @return double Array
- */
-static double *getKaKb() {
-    double *kAkB = malloc(sizeof(double) * 2);
-    if (kAkB == NULL) {
-        return NULL;
-    }
-    
-    kAkB[0] = (175.0 / 198.04) * (Yn + Xn);
-    kAkB[1] = (70.0 / 218.11) * (Yn + Zn);
-    
-    return kAkB;
-}
-
-Lab *getHunterLabFromXyz(Xyz *xyz) {
-    Lab *lab = malloc(sizeof(Lab));
-    if (lab == NULL) {
-        return NULL;
-    }
-    
-    if (xyz == NULL) {
-        lab->error = NULL_INPUT_PARAM;
-        return lab;
-    }
-    
-    double *kAkB = getKaKb();
-    if (kAkB == NULL) {
-        lab->error = MALLOC_ERROR;
-        return lab;
-    }
-    
-    lab->l = 100.0 * sqrt(xyz->y / Yn);
-    lab->a = 100.0 * kAkB[0] * ((xyz->x / Xn - xyz->y / Yn) / sqrtf(xyz->y / Yn));
-    lab->b = 100.0 * kAkB[1] * ((xyz->y / Yn - xyz->z / Zn) / sqrtf(xyz->y / Yn));
-    lab->error = NULL;
-    
-    free(kAkB);
-    
-    return lab;
-}
-
 /////// Xyz calculation
 
 /**
