+++ lymui/main.c
@@ -19,7 +19,6 @@ int main(int argc, const char * argv[]) {
     // Create suite case
     ctsuite_t *suite = ctsuite("Lymui test");
     ctcase_t  *srgb  = wrapSRgbCreationTest();
-    ctcase_t  *lab   = wrapLabCreationTest();
     ctcase_t  *luv   = wrapLuvCreationTest();
     ctcase_t  *lch   = wrapLchCreationTest();
     ctcase_t  *lchLab= wrapLchLabCreationTest();
@@ -29,7 +28,6 @@ int main(int argc, const char * argv[]) {
     
     // Launch suite
     ctscaseadd(suite, srgb);
-    ctscaseadd(suite, lab);
     ctscaseadd(suite, luv);
     ctscaseadd(suite, lch);
     ctscaseadd(suite, lchLab);
@@ -54,6 +52,7 @@ int main(int argc, const char * argv[]) {
     wrapXyzTest();
     wrapArgbTest();
     wrapLabTest();
+    wrapHLabTest();
     
     // free the suite
     ctsfree(suite);
